Create Ladder.cpp
Suppose there is a ladder with n steps.  A person standing at the  bottom of the  ladder and  wants to 
reach to its top. Assume the person can take a jump of size 1 to k at a time. Then, count the number 
of ways, the person can reach the top of the ladder.

Example 
	Input : n = 4, k = 3
	Output: 7
	Explanation : [1, 1, 1, 1], 
	              [1, 1, 2], 
	              [1, 2, 1], 
	              [2, 1, 1],
	              [2, 2],
	              [1, 3],
	              [3, 1]